You'll need to go through all of my JS pages again
I changed the format for reading settings data to match that of the weapons page and others by Stephen.  Should be no need to change anything in the home folder.  You will need to add the characterInfo (ObservableArray) and misc (Array) from the main-page-view-model and put that into the settings view model.

Other than making sure all the settings calls are updated the only thing that needs to be changed is adding the misc/character info data upload inside the navigation function for the main-page.  I did not add any additional files or folders :)